requirements.txt
openai
langchain
langchain-community
langgraph
python-dotenv